combine common regex code
